+++ scripts/mg_quickstart_test.py
@@ -7,9 +7,9 @@ async def main():
 
     # create a paperless client
     # we encourage you to use .env files to store your credentials later
-    url = "http://192.168.178.86:8100"
-    api_key = "d69b3da4dfd61dec7c60110bdae16637ced2b013"
-    async with PaperlessClient(url=url, api_key=api_key) as client:
+    url = "your url"
+    api_token = "your token"
+    async with PaperlessClient(url=url, api_token=api_token) as client:
         # List documents — full-text search across title and OCR content, return the last 3 docs
         docs = await client.documents.list(
             search="test", max_results=3, ordering="added", descending=True